:hammer: refactor mocks
