+++ src/resources/users/tests/users.service.spec.ts
@@ -3,21 +3,8 @@ import { getRepositoryToken } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from '../entities/user.entity';
 import { UsersService } from '../users.service';
-import { CreateUserDto } from '../dto/create-user.dto';
 import Bcrypt from 'src/shared/providers/hash/bcrypt';
-
-export const createUserMock: CreateUserDto = {
-  email: 'ok@email.com',
-  username: 'username',
-  password: '$2b$08$Fg1OiP6P3h4WEECv7a2Z2OdOLwxnW7FnINRg53J/W9QaJOhQCs4Oi',
-};
-
-export const userEntityMock: User = {
-  id: 'bab451d3-2833-411d-9e0a-fe7ae1cffbb1',
-  email: 'ok@email.com',
-  username: 'username',
-  password: '$2b$08$Fg1OiP6P3h4WEECv7a2Z2OdOLwxnW7FnINRg53J/W9QaJOhQCs4Oi',
-};
+import { createUserMock, userEntityMock } from './mocks/users_mocks';
 
 describe('UserService', () => {
   let service: UsersService;
